# src/app/modules/leave/leave.swagger.yaml
openapi: 3.0.0
info:
  title: Leave Request API
  description: API for creating, updating, deleting, and retrieving leave requests.
  version: 1.0.0
  contact:
    name: Your Company Name
    email: contact@company.com
    url: https://www.company.com
servers:
  - url: 'https://api.company.com/v1'
    description: Main production server

paths:
  /leave-requests:
    post:
      summary: Create a new leave request
      operationId: createLeaveRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: string
                  description: The employee's ID who is requesting leave
                  example: "607d1b2f3a7c1f19f3c30060"
                leaveType:
                  type: string
                  enum:
                    - Casual Leave
                    - Annual Leave
                  description: The type of leave being requested
                  example: "Casual Leave"
                fromDate:
                  type: string
                  format: date
                  description: The start date of the leave
                  example: "2025-02-01"
                toDate:
                  type: string
                  format: date
                  description: The end date of the leave
                  example: "2025-02-05"
                noOfDays:
                  type: number
                  description: The number of days of leave requested
                  example: 5
                remainingLeaves:
                  type: number
                  description: The number of remaining leaves available for the employee
                  example: 10
                reason:
                  type: string
                  description: The reason for the leave request
                  example: "Family emergency"
      responses:
        '200':
          description: Leave request successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Leave request created successfully"
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "607d1b2f3a7c1f19f3c30060"
                      employeeId:
                        type: string
                        example: "607d1b2f3a7c1f19f3c30060"
                      leaveType:
                        type: string
                        example: "Casual Leave"
                      fromDate:
                        type: string
                        format: date
                        example: "2025-02-01"
                      toDate:
                        type: string
                        format: date
                        example: "2025-02-05"
                      noOfDays:
                        type: number
                        example: 5
                      remainingLeaves:
                        type: number
                        example: 10
                      reason:
                        type: string
                        example: "Family emergency"
        '400':
          description: Invalid request data or missing fields
        '500':
          description: Internal server error

    get:
      summary: Retrieve all leave requests
      operationId: getAllLeaveRequests
      parameters:
        - name: page
          in: query
          description: The page number for pagination
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: The number of items per page
          required: false
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: List of leave requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Leave requests retrieved successfully"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          example: "607d1b2f3a7c1f19f3c30060"
                        employeeId:
                          type: string
                          example: "607d1b2f3a7c1f19f3c30060"
                        leaveType:
                          type: string
                          example: "Casual Leave"
                        fromDate:
                          type: string
                          format: date
                          example: "2025-02-01"
                        toDate:
                          type: string
                          format: date
                          example: "2025-02-05"
                        noOfDays:
                          type: number
                          example: 5
                        remainingLeaves:
                          type: number
                          example: 10
                        reason:
                          type: string
                          example: "Family emergency"
        '500':
          description: Internal server error

  /leave-requests/{leaveRequestsId}:
    get:
      summary: Retrieve a single leave request by ID
      operationId: getSingleLeaveRequest
      parameters:
        - name: leaveRequestsId
          in: path
          required: true
          description: The ID of the leave request to retrieve
          schema:
            type: string
            example: "607d1b2f3a7c1f19f3c30060"
      responses:
        '200':
          description: Leave request successfully retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Leave request retrieved successfully"
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "607d1b2f3a7c1f19f3c30060"
                      employeeId:
                        type: string
                        example: "607d1b2f3a7c1f19f3c30060"
                      leaveType:
                        type: string
                        example: "Casual Leave"
                      fromDate:
                        type: string
                        format: date
                        example: "2025-02-01"
                      toDate:
                        type: string
                        format: date
                        example: "2025-02-05"
                      noOfDays:
                        type: number
                        example: 5
                      remainingLeaves:
                        type: number
                        example: 10
                      reason:
                        type: string
                        example: "Family emergency"
        '400':
          description: Invalid leave request ID
        '404':
          description: Leave request not found
        '500':
          description: Internal server error

    patch:
      summary: Update a leave request by ID
      operationId: updateLeaveRequest
      parameters:
        - name: leaveRequestsId
          in: path
          required: true
          description: The ID of the leave request to update
          schema:
            type: string
            example: "607d1b2f3a7c1f19f3c30060"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  type: string
                  description: The employee's ID who is requesting leave
                  example: "607d1b2f3a7c1f19f3c30060"
                leaveType:
                  type: string
                  enum:
                    - Casual Leave
                    - Annual Leave
                  description: The type of leave being requested
                  example: "Casual Leave"
                fromDate:
                  type: string
                  format: date
                  description: The start date of the leave
                  example: "2025-02-01"
                toDate:
                  type: string
                  format: date
                  description: The end date of the leave
                  example: "2025-02-05"
                noOfDays:
                  type: number
                  description: The number of days of leave requested
                  example: 5
                remainingLeaves:
                  type: number
                  description: The number of remaining leaves available for the employee
                  example: 10
                reason:
                  type: string
                  description: The reason for the leave request
                  example: "Family emergency"
      responses:
        '200':
          description: Leave request successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Leave request updated successfully"
                  data:
                    type: object
                    properties:
                      _id:
                        type: string
                        example: "607d1b2f3a7c1f19f3c30060"
        '400':
          description: Invalid leave request ID or data
        '404':
          description: Leave request not found
        '500':
          description: Internal server error

    delete:
      summary: Delete a leave request by ID
      operationId: deleteLeaveRequest
      parameters:
        - name: leaveRequestsId
          in: path
          required: true
          description: The ID of the leave request to delete
          schema:
            type: string
            example: "607d1b2f3a7c1f19f3c30060"
      responses:
        '200':
          description: Leave request successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Leave request deleted successfully"
        '400':
          description: Invalid leave request ID
        '404':
          description: Leave request not found
        '500':
          description: Internal server error